+++ js/data.js
@@ -1,0 +1,53 @@
+const DESCRIPTIONS = [
+  'Красивое место',
+  'I Love U',
+  'Where i am?',
+  'Taste food',
+  'Beautiful sound'
+];
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const USER_NAMES = [
+  'Анастасия',
+  'Елизавета',
+  'Ольга',
+  'Алексей',
+  'Семён'
+];
+
+const POST_COUNT = 25;
+
+const RangeIDs = {
+  MIN: 1,
+  MAX: 25
+};
+
+const RangeIDComment = {
+  MIN: 1,
+  MAX: 1000
+};
+
+const RangeURLs = {
+  MIN: 1,
+  MAX: 25
+};
+
+const RangeLikes = {
+  MIN: 15,
+  MAX: 200
+};
+
+const countOfComments = {
+  MIN: 0,
+  MAX: 25
+};
+
+export {DESCRIPTIONS, COMMENT_MESSAGES, USER_NAMES, POST_COUNT, RangeIDs, RangeIDComment, RangeURLs, RangeLikes, countOfComments};
